queue: move to own file
