+++ blink/blink.emProject
@@ -19,6 +19,9 @@
     <file file_name="scan.c" />
     <file file_name="scan.h" />
 
+    <file file_name="queue.c" />
+    <file file_name="queue.h" />
+
     <file file_name="scheduler.c" />
     <file file_name="all_schedules.c" />
     <file file_name="scheduler.h" />
